app/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-jose
passlib
python-multipart
openai
requests
alembic
python-dotenv
pytest
httpx
bcrypt
pydantic[email]
langfuse